Restructuring of consensus components. Start on finalization processing.
